Added concrete refactoring directory
