+++ src/ecooptimizer/refactorers/concrete/long_element_chain.py
@@ -4,8 +4,8 @@ from pathlib import Path
 import re
 from typing import Any, Optional
 
-from .multi_file_refactorer import MultiFileRefactorer
-from ..data_types.smell import LECSmell
+from ..multi_file_refactorer import MultiFileRefactorer
+from ...data_types.smell import LECSmell
 
 
 class DictAccess:
