+++ src/ecooptimizer/refactorers/concrete/long_lambda_function.py
@@ -1,7 +1,7 @@
 from pathlib import Path
 import re
-from .base_refactorer import BaseRefactorer
-from ..data_types.smell import LLESmell
+from ..base_refactorer import BaseRefactorer
+from ...data_types.smell import LLESmell
 
 
 class LongLambdaFunctionRefactorer(BaseRefactorer[LLESmell]):
