+++ tests/smells/test_long_lambda_function.py
@@ -4,7 +4,7 @@ import pytest
 
 from ecooptimizer.analyzers.analyzer_controller import AnalyzerController
 from ecooptimizer.data_types.smell import LLESmell
-from ecooptimizer.refactorers.long_lambda_function import LongLambdaFunctionRefactorer
+from ecooptimizer.refactorers.concrete.long_lambda_function import LongLambdaFunctionRefactorer
 from ecooptimizer.utils.smell_enums import CustomSmell
